feature:setting 모듈의 View 관련 코드 제거 (#137)
* #136 - remove : setting의 view 코드들 삭제

* #136 - style : screen에 있던 코드들을 setting 패키지 하위로 이동
