+++ feature/main/src/main/java/com/example/main/app/RoutinerApp.kt
@@ -22,7 +22,7 @@ import com.hegunhee.repeat.navigation.repeatRoutineNavGraph
 import com.hegunhee.routiner.insertRoutine.screen.navigation.INSERT_REPEAT_ROUTINE_ROUTE
 import com.hegunhee.routiner.insertRoutine.screen.navigation.INSERT_ROUTINE_ROUTE
 import com.hegunhee.routiner.insertRoutine.screen.navigation.insertNavGraph
-import com.hegunhee.setting.screen.navigation.settingNavGraph
+import com.hegunhee.setting.navigation.settingNavGraph
 import kotlinx.coroutines.CoroutineScope
 
 @Composable
