+++ feature/setting/src/main/java/com/hegunhee/setting/navigation/SettingNavigation.kt
@@ -1,8 +1,8 @@
-package com.hegunhee.setting.screen.navigation
+package com.hegunhee.setting.navigation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.hegunhee.setting.screen.SettingRootScreen
+import com.hegunhee.setting.SettingRootScreen
 
 const val SETTING_ROUTE = "SETTING"
