--- feature/setting/src/main/res/layout/fragment_setting.xml
@@ -1,127 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<layout>
-
-    <data>
-
-        <variable
-            name="viewModel"
-            type="com.hegunhee.setting.SettingViewModel" />
-
-        <import type="android.view.View"/>
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <TextView
-            android:id="@+id/alarmSettingHeader"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:text="Alarm Settings"
-            android:textColor="@color/white"
-            android:textSize="30sp"
-            android:textStyle="bold"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/alarmSettingLayout"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="20dp"
-            android:layout_marginTop="10dp"
-            android:background="@drawable/rounded_shape_gray_20dp"
-            android:backgroundTint="@color/gray"
-            android:padding="10dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/alarmSettingHeader">
-
-            <TextView
-                android:id="@+id/enableAlarmHeader"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="알람 설정"
-                android:textColor="@color/white"
-                android:textSize="20sp"
-                app:layout_constraintBottom_toBottomOf="@id/alarmEnableSwtich"
-                app:layout_constraintStart_toStartOf="@id/alarmSettingLayout"
-                app:layout_constraintTop_toTopOf="@id/alarmEnableSwtich" />
-
-            <androidx.appcompat.widget.SwitchCompat
-                android:id="@+id/alarmEnableSwtich"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:checked="@{viewModel.alarmTime.enableAlarm}"
-                android:onCheckedChanged="@{(switch, checked) -> viewModel.onEnableAlarmSwitchClick(checked)}"
-                app:layout_constraintEnd_toEndOf="@id/alarmSettingLayout"
-                app:layout_constraintTop_toTopOf="@id/alarmSettingLayout" />
-
-            <TextView
-                android:id="@+id/alarmTimeHeader"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="시간 설정"
-                android:textColor="@color/white"
-                android:textSize="20sp"
-                android:visibility="@{viewModel.alarmTime.enableAlarm ? View.VISIBLE : View.INVISIBLE}"
-                app:layout_constraintBottom_toBottomOf="@id/minuteSpinner"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="@id/minuteSpinner" />
-
-            <androidx.appcompat.widget.AppCompatSpinner
-                android:id="@+id/hourSpinner"
-                android:layout_width="wrap_content"
-                android:layout_height="50dp"
-                app:layout_constraintBottom_toBottomOf="@id/minuteSpinner"
-                app:layout_constraintEnd_toStartOf="@id/alarmColonText"
-                android:visibility="@{viewModel.alarmTime.enableAlarm ? View.VISIBLE : View.INVISIBLE}"
-                app:layout_constraintTop_toTopOf="@id/minuteSpinner" />
-
-            <TextView
-                android:id="@+id/alarmColonText"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text=":"
-                android:textColor="@color/white"
-                android:textSize="20sp"
-                android:textStyle="bold"
-                app:layout_constraintBottom_toBottomOf="@id/minuteSpinner"
-                android:visibility="@{viewModel.alarmTime.enableAlarm ? View.VISIBLE : View.INVISIBLE}"
-                app:layout_constraintEnd_toStartOf="@id/minuteSpinner"
-                app:layout_constraintTop_toTopOf="@id/minuteSpinner" />
-
-
-            <androidx.appcompat.widget.AppCompatSpinner
-                android:id="@+id/minuteSpinner"
-                android:layout_width="wrap_content"
-                android:layout_height="50dp"
-                android:layout_marginTop="10dp"
-                app:layout_constraintEnd_toEndOf="@id/alarmSettingLayout"
-                android:visibility="@{viewModel.alarmTime.enableAlarm ? View.VISIBLE : View.INVISIBLE}"
-                app:layout_constraintTop_toBottomOf="@id/alarmEnableSwtich" />
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/alarmSaveButton"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:background="@drawable/rounded_shape_gray_20dp"
-                android:backgroundTint="@color/sky_blue"
-                android:text="저장하기"
-                android:textColor="@color/black"
-                android:visibility="@{viewModel.alarmTime.enableAlarm ? View.VISIBLE : View.INVISIBLE}"
-                android:textSize="20sp"
-                android:onClick="@{() -> viewModel.onSaveAlarmClick()}"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/alarmTimeHeader" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</layout>
