--- feature/setting/src/main/res/layout/spinner_item.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<TextView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:textSize="13sp"
-    android:paddingStart="10dp"
-    android:textColor="@color/white"/>
